sieve of eratosthenes and many more

+++ PracticeQuestions/p2.py
@@ -2,6 +2,6 @@
 def add(a,b):
     return a+b
 
-a = int(input("enter number 1:"))
-b = int(input("enter number 2: "))
+a = float(input("enter number 1:"))
+b = float(input("enter number 2: "))
 print(f"sum = {add(a,b)}")
